Move download method  to DownloadService from Notification
